+++ core/CHANGELOG.md
@@ -15,3 +15,9 @@
 *   Ensure transition to payment processing state happens outside transaction.
 
     Chris Salzberg
+
+* Refactored Stock::Coordinator to optionally accept a list of inventory units 
+  for an order so that shipments can be created for an order that do not comprise
+  only of the order's line items.
+
+     Andrew Thal